Minor tweaks to GCClassic integration tests
test/shared/commonFunctionsForTests.sh
- Ask if the user REALLY wants to delete leftover files in the
  integration test root folder before proceeding.  If not, the
  return error code 1.

test/GCClassic/intTestCreate.sh:
- Now name the run directories more consistently with the default
  run directory names that you get from the createRunDir.sh script
  (e.g. gc_4x5_fullchem_benchmark_merra2).  We have to tag the
  met field name on the end of the run directory name to keep them
  separate.

test/GCClassic/intTest_slurm.sh:
test/GCClassic/intTest_lsf.sh
- Put an error check to trap the case of the user saying that they
  do not want to delete leftover files from the integration test
  root folder.  If this happens, exit without doing anything.

test/intTestExecute_slurm.sh
- Increase run time to 16 hours, just in case there are slowdowns.

+++ test/shared/commonFunctionsForTests.sh
@@ -167,14 +167,26 @@ function cleanup_files() {
     # 1st argument = root folder for tests (w/ many rundirs etc)
     #========================================================================
     if [[ "x${1}" != "x" ]]; then
-	printf "Removing leftover run directories and scripts:\n"
+
+	# Give user a chance to avoid removing files
+	printf "\nRemoving files and directories in ${1}:\n"
+	printf "If this is OK, type 'yes to proceed or 'no' to quit'\n"
+	read answer
+	if [[ ! ${answer} =~ [Yy][Ee][Ss] ]]; then
+	    printf "Exiting..."
+	    return 1
+	fi
+
+	# Remove files and unlink links
 	for file in ${1}/*; do
 	    if [[ -h ${file} ]]; then
 		unlink ${file}
 	    else
 		path=$(absolute_path ${file})
-		printf " ... ${path}\n";
-		rm -rf ${path}
+		if [[ -e ${path} ]]; then
+		    printf " ... ${path}\n";
+		    rm -rf ${path}
+		fi
 		unset path
 	    fi
 	done
